+++ test/unit/processes.ts
@@ -97,8 +97,8 @@ describe("Governance Process", () => {
             expect(data.envelopeType.value).to.equal(DEFAULT_PROCESS_MODE)
             expect(data.censusOrigin.value).to.eq(DEFAULT_CENSUS_ORIGIN)
             expect(data.metadata.toLowerCase()).to.equal(DEFAULT_METADATA_CONTENT_HASHED_URI)
-            expect(data.censusMerkleRoot).to.equal(DEFAULT_MERKLE_ROOT)
-            expect(data.censusMerkleTree).to.equal(DEFAULT_MERKLE_TREE_CONTENT_HASHED_URI)
+            expect(data.censusRoot).to.equal(DEFAULT_MERKLE_ROOT)
+            expect(data.censusUri).to.equal(DEFAULT_MERKLE_TREE_CONTENT_HASHED_URI)
             expect(data.entityAddress).to.equal(entityAccount.address)
             expect(data.status.isPaused).to.eq(true)
             expect(data.maxCount).to.eq(DEFAULT_MAX_COUNT)
@@ -217,7 +217,7 @@ describe("Governance Process", () => {
             let processId = "0x8b35e10045faa886bd2e18636cd3cb72e80203a04e568c47205bf0313a0f60d1"
             let siblings = "0x0003000000000000000000000000000000000000000000000000000000000006f0d72fbd8b3a637488107b0d8055410180ec017a4d76dbb97bee1c3086a25e25b1a6134dbd323c420d6fc2ac3aaf8fff5f9ac5bc0be5949be64b7cfd1bcc5f1f"
 
-            const { envelope: e1, signature: signature1 } = await VotingApi.packageSignedEnvelope({ censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN), votes: [1, 2, 3], censusProof: siblings, processId, walletOrSigner: wallet })
+            const { envelope: e1, signature: signature1 } = await VotingApi.packageSignedEnvelope({ censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN_TREE), votes: [1, 2, 3], censusProof: siblings, processId, walletOrSigner: wallet })
             const envelope1 = VoteEnvelope.deserializeBinary(e1)
             expect(Buffer.from(envelope1.getProcessid()).toString("hex")).to.eq(processId.slice(2))
             expect(Buffer.from(envelope1.getProof().getGraviton().getSiblings()).toString("hex")).to.eq(siblings.slice(2))
@@ -229,7 +229,7 @@ describe("Governance Process", () => {
             processId = "0x36c886bd2e18605bf03a0428be100313a0f6e568c470d135d3cb72e802045faa"
             siblings = "0x0003000000100000000002000000000300000000000400000000000050000006f0d72fbd8b3a637488107b0d8055410180ec017a4d76dbb97bee1c3086a25e25b1a6134dbd323c420d6fc2ac3aaf8fff5f9ac5bc0be5949be64b7cfd1bcc5f1f"
 
-            const { envelope: e2, signature: signature2 } = await VotingApi.packageSignedEnvelope({ censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN), votes: [5, 6, 7], censusProof: siblings, processId, walletOrSigner: wallet })
+            const { envelope: e2, signature: signature2 } = await VotingApi.packageSignedEnvelope({ censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN_TREE), votes: [5, 6, 7], censusProof: siblings, processId, walletOrSigner: wallet })
             const envelope2 = VoteEnvelope.deserializeBinary(e2)
             expect(Buffer.from(envelope2.getProcessid()).toString("hex")).to.eq(processId.slice(2))
             expect(Buffer.from(envelope2.getProof().getGraviton().getSiblings()).toString("hex")).to.eq(siblings.slice(2))
@@ -239,7 +239,7 @@ describe("Governance Process", () => {
             expect(BytesSignature.isValid(signature2, wallet._signingKey().publicKey, e2)).to.eq(true)
 
             expect(async () => {
-                await VotingApi.packageSignedEnvelope({ censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN), votes: ["1", "2", "3"], censusProof: siblings, processId, walletOrSigner: wallet } as any)
+                await VotingApi.packageSignedEnvelope({ censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN_TREE), votes: ["1", "2", "3"], censusProof: siblings, processId, walletOrSigner: wallet } as any)
             }).to.throw
         })
 
@@ -270,7 +270,7 @@ describe("Governance Process", () => {
             // one key
             for (let item of processes) {
                 const processKeys = { encryptionPubKeys: [{ idx: 1, key: votePublicKey }] }
-                const { envelope: envelopeBytes, signature } = await VotingApi.packageSignedEnvelope({ censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN), votes: item.votes, censusProof: item.siblings, processId: item.processId, walletOrSigner: wallet, processKeys })
+                const { envelope: envelopeBytes, signature } = await VotingApi.packageSignedEnvelope({ censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN_TREE), votes: item.votes, censusProof: item.siblings, processId: item.processId, walletOrSigner: wallet, processKeys })
                 const envelope = VoteEnvelope.deserializeBinary(envelopeBytes)
 
                 expect(Buffer.from(envelope.getProcessid()).toString("hex")).to.eq(item.processId.slice(2))
@@ -330,7 +330,7 @@ describe("Governance Process", () => {
                 const processKeys = { encryptionPubKeys: encryptionKeys.map((kp, idx) => ({ idx, key: kp.publicKey })) }
 
                 const { envelope: envelopeBytes, signature } = await VotingApi.packageSignedEnvelope({
-                    censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN),
+                    censusOrigin: new ProcessCensusOrigin(ProcessCensusOrigin.OFF_CHAIN_TREE),
                     votes: item.votes,
                     censusProof: item.siblings,
                     processId: item.processId,